Renamed eq_featureoperation_fixture -> eq_featureoperation_combs Moved and renamed featureoperation_combos.py -> featureoperation_util.py Extracted DataFrameMock from unitutil.py to dataframewithinfo_util.py
